% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignMapColor.R
\name{assignMapColor}
\alias{assignMapColor}
\title{Assign Country Color}
\usage{
assignMapColor(pisaR, country, color)
}
\arguments{
\item{country}{}

\item{color}{}
}
\value{
pisaR htmlwidget object
}
\description{
Allows User to assign color to a specific country
}
