% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-initialize.R
\name{createISAtab}
\alias{createISAtab}
\title{Export to ISA-tab}
\usage{
createISAtab(iroot, files, path, zip = TRUE, isamap)
}
\arguments{
\item{iroot}{path to investigation. If missing, a interactively navigate
to the desired investigation and select investigation metadata
file.}

\item{files}{a vector of assay metadata file names
(with study and assay path) to export. If missing,
all assays in selected investigation will be processed.}

\item{path}{path to the exported files, default is /ISA-tab/ folder in the
investigation path.}

\item{zip}{controls creation of zip file, created if TRUE.}

\item{isamap}{path and name of the file with mapping of
pISA-tree and ISA-tab keys.
If missing, the mapping file that comes with
the package pisar is used.}
}
\value{
character vector, names of created files.
}
\description{
Export pISA-tree investigation into ISA-tab files.
}
\examples{
\dontrun{
# interactively select investigation metadata file for export.
isafiles <- createISAtab()
# export one assay from the demonstrative investigation
iroot <- file.path(find.package("pisar"),"extdata/_p_Test-pISA-tab/_I_i")
files <- "_S_s/_A_a-GCMS/_ASSAY_METADATA.TXT"
path <- "./ISA-tab"
createISAtab(iroot, files, path)
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{ISA-tab}
\keyword{pISA-tree,}
