% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLayer.R
\name{createLayer}
\alias{createLayer}
\title{Create Plot Layer}
\usage{
createLayer(pisaR, layerType, layerColor, layerLabel, layerData = NULL,
  layerMapping)
}
\arguments{
\item{layerType}{defines the layer as either a 'heatmap' or 'globalMap' (anticipates future layer types)}

\item{layerColor}{a string or vector of strings of HTML recognized colors}

\item{layerLabel}{a string unique to the layer for the plot}

\item{layerData}{(optional) A data frame or tibble}

\item{layerMapping}{a list object of variables defined, eg. list(x_var = "xVar", y_var = "yVar", z_var = "colorVar")}
}
\value{
A list object, pisaR, with layers now attached
}
\description{
Create plot layer by defining parameters
}
