% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineColorScale.R
\name{defineColorScale}
\alias{defineColorScale}
\title{Define Color Scale}
\usage{
defineColorScale(pisaR, color_palette, key)
}
\arguments{
\item{color_palette}{a list of character strings defining HTML recognized colors}

\item{key}{a list of character strings mapping color to a data element}
}
\value{
pisaR htmlwidget object
}
\description{
Allows User Defined ordinial color scale
}
