% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definePlotMargin.R
\name{definePlotMargin}
\alias{definePlotMargin}
\title{Define Plot Margins}
\usage{
definePlotMargin(pisaR, top = 50, left = 50, bottom = 0, right = 50)
}
\arguments{
\item{top, left, bottom, right}{numeric values representing respective plot margins}
}
\value{
pisaR htmlwidget object
}
\description{
Allows User Defined plot margins
}
