% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineTimeInterval.R
\name{defineTimeInterval}
\alias{defineTimeInterval}
\title{Define Time Interval}
\usage{
defineTimeInterval(pisaR, interval)
}
\arguments{
\item{intervals}{vector of date/character values}
}
\value{
pisaR htmlwidget object
}
\description{
Allows User to define the time initerval for the x axis
}
