% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-api.R
\name{fhCreate}
\alias{fhCreate}
\title{Create pISA layer or *fh* component.}
\usage{
fhCreate(type = "assays", meta = list(), class = "EXP",
  file = "NA.TXT")
}
\arguments{
\item{type}{Component name (e.g. 'people', 'projets', ...).}

\item{meta}{Data frame with pISA metadata or
a list with minimal information (Title, 
Description, *ToDo: add fields*).}

\item{class}{Assay class key string.
Possible values are 'EXP' and 'MODEL'.}

\item{file}{File name with path, relative to layer.}
}
\value{
FAIRDOMhub created component.
}
\description{
Create pISA layer or *fh* component.
}
\note{
Upon success (status code 200) details
of newly created component can be used. Check status code.
}
\examples{
\\donotrun{
if(FALSE)
{
fhIni(prid = 26, test=TRUE)
options("fhub")
 sp <- fhCreate( type = "projects"
  , meta= list(
      Title=paste("Test project", Sys.time())
    , Description="Testing of upload")
    )
 str(sp)
 str(fhData(sp))
# Add member manually
 fhIni(prid = 26, pid=104, test=TRUE)
 options("fhub")
 si <- fhCreate( type = "investigations"
  , meta= list(
      Title=paste("Test investigation", Sys.time())
    , Description="Testing of upload")
    )
 si
 fhData(si)$id

 iid=fhData(si)$id
 iid <- 115
 fhIni(prid = 26, pid=104, iid=iid, test=TRUE)
 options("fhub")
 ss <- fhCreate( type = "studies"
  , meta= list(
      Title=paste("Test study", Sys.time())
    , Description="Testing of upload")
    )
 ss
 fhData(ss)$id

 fhIni(prid = 26
       , pid=104
       , iid=fhData(si)$id
       , sid=fhData(ss)$id
       , test=TRUE)
 options("fhub")
 sa <- fhCreate( type = "assays"
  , meta= list(
      Title=paste("Test assay", Sys.time())
    , Description="Testing of upload")
    , class="EXP"
    )
 #str(sa)
 sa
}

# Type: data_file
astring <- "_p_Demo/_I_Test/_S_Show/_A_Work-R/"
oldwd <- setwd(system.file("extdata",astring,package="pisar"))
oldwd
.aname <- getLayer("A")
.aroot <- getRoot("A")
.ameta  <- readMeta()
 file <- "input/README.MD"
 fhIni(prid = 26, pid=104, iid=115 , sid=117 , aid=401, test=TRUE)
 type <- "data_files"
 type <- "documents"
 sdat <- fhCreate( type = type
  , meta= list(
      Title=paste("Test assay", Sys.time())
    , Description="Testing of upload")
  , file=file
    )
 #str(sdat)
 sdat
 fhData(sdat)$id
if(interactive()) setwd(oldwd)
getwd()
res <- sdat$content
item_link <- file.path(res$meta$base_url,res$links$self)
if(interactive()) shell.exec(item_link) 
}
}
\seealso{
\code{\link{fhGet}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
