% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-api.R
\name{fhFindTitle}
\alias{fhFindTitle}
\title{Get details of component with id from an *fh* object.}
\usage{
fhFindTitle(type, id)
}
\arguments{
\item{type}{Components name (e.g. 'people', 'projets', ...).}

\item{id}{Character string with the identifier
of the component (id part).}
}
\value{
FAIRDOMhub component identifier: id, type and title.
    If argument title is missing, a data frame with identifiers
    for all items is returned. See note.
}
\description{
Get details of component with id from an *fh* object.
}
\note{
If item is not found, empty string is returned as title.
}
\examples{
\\donotrun{
fhIni()
id <- fhFindId("people","Guest")
id
idNew <- fhFindTitle("people", id[1])
idNew
all.equal(id, idNew)
fhFindTitle("people",0)
}
}
\seealso{
\code{\link{fhFindTitle}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
