% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-api.R
\name{fhLog}
\alias{fhLog}
\title{Writes a note to a log file.}
\usage{
fhLog(..., file = "FAIRDOM.log", append = TRUE)
}
\arguments{
\item{...}{Objects to form a line.}

\item{file}{Log file name.}

\item{append}{Control append/rewrite mode.}
}
\value{
System time of invoking..
}
\description{
Writes a note to a log file.
}
\examples{
tst <- function(){
  fhLog("Test", "writing to logfile", file="")
  fht <- system.time(Sys.sleep(1))
  fhLog( "Time:", round(fht["elapsed"],2))
}
tst()
rm(tst)

}
\seealso{
\code{\link{fhGet}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
