% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-api.R
\name{fhSkeleton}
\alias{fhSkeleton}
\title{Create *fh* skeleton.}
\usage{
fhSkeleton(type = "assay", meta, file)
}
\arguments{
\item{type}{Component name (e.g. 'people', 'projets', ...).}

\item{meta}{Data frame with pISA metadata or
a list with minimal information (Title, Description, *ToDo: add fields*).}
}
\value{
A list with the minimal information structure.
}
\description{
Creates *fh* object with required structure.
}
\examples{
\\donotrun{
require(jsonlite)
meta= list(Title = "Test layer", Description = "Some description")
type = "projects"
sp <- fhSkeleton( type = type
  , meta= meta
  )
str(sp)
type = "investigations"
si <- fhSkeleton( type = type
  , meta= meta
  )
str(si)
type = "studies"
ss <- fhSkeleton( type = type
  , meta= meta
  )
str(ss)
type = "assays"
sa <- fhSkeleton( type = type
  , meta= meta
  )
str(sa)

type = "data_files"
file =
sdata <- fhSkeleton( type = type
  , meta= meta
  )
str(sdata)

}
}
\seealso{
\code{\link{fhCreate}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
