% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-api.R
\name{fhUpload}
\alias{fhUpload}
\title{Upload file.}
\usage{
fhUpload(object, file)
}
\arguments{
\item{object}{A previously created SEEK component.
Must be one of 'documents' or 'data_files'.}

\item{file}{File name with path, relative to layer.}
}
\value{
FAIRDOMhub created component.
}
\description{
Upload file to a crerated object of type 'documents' or 'data_files'.
}
\note{
Upon success (status code 200) details
of newly created component can be used. Check status code.
}
\examples{
 iid <- 115
 fhIni(prid = 26, pid=104, iid=iid, aid= test=TRUE)
meta <- list(Title="Title", Description="Description")
file <- dir()[1]
file
fhd <- fhCreate("documents",meta,file=file)
fhd
sda <- fhUpload(fhd, file)

}
\seealso{
\code{\link{fhCreate}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
