% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-initialize.R
\name{getMeta}
\alias{getMeta}
\title{Get metadata value}
\usage{
getMeta(x, key, nl = TRUE)
}
\arguments{
\item{x}{Two column character data frame with Key / Value pairs.}

\item{key}{String, key name. Trailing colon can be omitted.}

\item{nl}{Logical, expand backslash character for new lines.}

\item{...}{Any other arguments (not used at the moment).}
}
\value{
Character string with key value.
}
\description{
Get metadata value
}
\note{
Parameter item is matched exactly to the item names.
}
\examples{
\dontrun{
astring <- "_p_Demo/_I_Test/_S_Show/_A_Work-R/other"
oldwd <- setwd(system.file("extdata",astring,package="pisar"))
oldwd
.iroot <- getRoot("I")
.imeta <- readMeta(.iroot)
getMeta(.imeta, "Description")
# list
listmeta <- list(Title = "My title"
   , Description = "A longer description")
getMeta( listmeta, "Title")
getMeta( listmeta, "Description")
setwd(oldwd)
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
