% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-initialize.R
\name{pasteMeta}
\alias{pasteMeta}
\alias{pasteMeta.default}
\alias{pasteMeta.list}
\title{Paste metadata values}
\usage{
pasteMeta(x, kvsep = "\\t", nlsep = "\\n")

\method{pasteMeta}{default}(x, kvsep = "\\t", nlsep = "\\n")

\method{pasteMeta}{list}(x, kvsep = ":\\t", nlsep = "\\n")
}
\arguments{
\item{x}{Two column character data frame or with Key / Value pairs or
a list of Vales named by Keys.}

\item{kvsep}{Key/Value separator (default is tab character).}

\item{nlsep}{line separator (default is new line).}
}
\value{
Character string with concatenated key values.
}
\description{
Paste metadata values into a strig.
}
\examples{
\dontrun{
astring <- "_p_Demo/_I_Test/_S_Show/_A_Work-R/other"
oldwd <- setwd(system.file("extdata",astring,package="pisar"))
oldwd
.iroot <- getRoot("I")
.imeta <- readMeta(.iroot)
x <- pasteMeta(.imeta)
x
cat(x)
# list
listmeta <- list(Title = "My title"
   , Description = "A longer description")
x <- pasteMeta( listmeta)
x
cat(x)
setwd(oldwd)
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
