% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisaR.R
\name{pisaR}
\alias{pisaR}
\title{pisaR}
\usage{
pisaR(data = NULL, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{(optional) A data frame or tibble}
}
\description{
Widget for creating PISA application pieces
}
