% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-initialize.R
\docType{package}
\name{pisar}
\alias{pisar}
\title{pisar: pISA-tree support functions}
\description{
The package provides several functions for support
and use of pISA-tree.
}
\section{Usage}{

pISA-tree is a standardized directory tree
for storing projet information under ISA paradigm.
The set of functions have two fold purpose:
\enumerate{
    \item To enable use of metadata for reproducible documents.
    \item To enable automated upload to external repository.
(e. g. FAIRDOMhub).}
}

