% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-initialize.R
\docType{package}
\name{pisar}
\alias{pisar}
\title{pisar: pISA-tree support functions}
\description{
The package provides several functions for support
and use of pISA-tree.
}
\section{Usage}{

pISA-tree is a standardized directory tree
for storing project information under ISA paradigm.
This set of functions 
enables use of metadata for reproducible documents.
}

